services/mcp_server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.3
beautifulsoup4==4.13.4
python-dateutil==2.9.0.post0
playwright==1.55.0
